+++ bindings/python/src/lib.rs
@@ -67,7 +67,7 @@ fn parse_url(url: Option<String>) -> PyResult<Option<url::Url>> {
     })
 }
 
-/// CSSInliner(inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None, preallocate_node_capacity=0)
+/// CSSInliner(inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None, preallocate_node_capacity=0)
 ///
 /// Customizable CSS inliner.
 #[pyclass]
@@ -78,7 +78,7 @@ struct CSSInliner {
 #[pymethods]
 impl CSSInliner {
     #[new(
-        text_signature = "(inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None, preallocate_node_capacity=0)"
+        text_signature = "(inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None, preallocate_node_capacity=0)"
     )]
     fn new(
         inline_style_tags: Option<bool>,
@@ -90,7 +90,7 @@ impl CSSInliner {
     ) -> PyResult<Self> {
         let options = rust_inline::InlineOptions {
             inline_style_tags: inline_style_tags.unwrap_or(true),
-            remove_style_tags: remove_style_tags.unwrap_or(false),
+            remove_style_tags: remove_style_tags.unwrap_or(true),
             base_url: parse_url(base_url)?,
             load_remote_stylesheets: load_remote_stylesheets.unwrap_or(true),
             extra_css: extra_css.map(Cow::Owned),
@@ -119,12 +119,12 @@ impl CSSInliner {
     }
 }
 
-/// inline(html, inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None)
+/// inline(html, inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None)
 ///
 /// Inline CSS in the given HTML document
 #[pyfunction]
 #[pyo3(
-    text_signature = "(html, inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None)"
+    text_signature = "(html, inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None)"
 )]
 fn inline(
     html: &str,
@@ -137,7 +137,7 @@ fn inline(
 ) -> PyResult<String> {
     let options = rust_inline::InlineOptions {
         inline_style_tags: inline_style_tags.unwrap_or(true),
-        remove_style_tags: remove_style_tags.unwrap_or(false),
+        remove_style_tags: remove_style_tags.unwrap_or(true),
         base_url: parse_url(base_url)?,
         load_remote_stylesheets: load_remote_stylesheets.unwrap_or(true),
         extra_css: extra_css.map(Cow::Borrowed),
@@ -148,12 +148,12 @@ fn inline(
     Ok(inliner.inline(html).map_err(InlineErrorWrapper)?)
 }
 
-/// inline_many(htmls, inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None)
+/// inline_many(htmls, inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None)
 ///
 /// Inline CSS in multiple HTML documents
 #[pyfunction]
 #[pyo3(
-    text_signature = "(htmls, inline_style_tags=True, remove_style_tags=False, base_url=None, load_remote_stylesheets=True, extra_css=None)"
+    text_signature = "(htmls, inline_style_tags=True, remove_style_tags=True, base_url=None, load_remote_stylesheets=True, extra_css=None)"
 )]
 fn inline_many(
     htmls: &PyList,
@@ -166,7 +166,7 @@ fn inline_many(
 ) -> PyResult<Vec<String>> {
     let options = rust_inline::InlineOptions {
         inline_style_tags: inline_style_tags.unwrap_or(true),
-        remove_style_tags: remove_style_tags.unwrap_or(false),
+        remove_style_tags: remove_style_tags.unwrap_or(true),
         base_url: parse_url(base_url)?,
         load_remote_stylesheets: load_remote_stylesheets.unwrap_or(true),
         extra_css: extra_css.map(Cow::Borrowed),