+++ css-inline/tests/test_inlining.rs
@@ -139,14 +139,8 @@ fn simple_merge() {
     let html = html!(style, r#"<h1 style="font-size: 1px">Big Text</h1>"#);
     let inlined = inline(&html).unwrap();
     // Then new styles should be merged with the existing ones
-    let option_1 = html!(
-        style,
-        r#"<h1 style="font-size: 1px;color: red">Big Text</h1>"#
-    );
-    let option_2 = html!(
-        style,
-        r#"<h1 style="color: red;font-size: 1px">Big Text</h1>"#
-    );
+    let option_1 = html!(r#"<h1 style="font-size: 1px;color: red">Big Text</h1>"#);
+    let option_2 = html!(r#"<h1 style="color: red;font-size: 1px">Big Text</h1>"#);
     let valid = (inlined == option_1) || (inlined == option_2);
     assert!(valid, "{}", inlined);
 }
@@ -211,7 +205,7 @@ fn href_attribute_unchanged() {
         inlined,
         r#"<html><head>
     <title>Test</title>
-    <style>h1 { color:blue; }</style>
+    
 </head>
 <body>
     <h1 style="color:blue;">Big Text</h1>
@@ -250,20 +244,12 @@ fn complex_child_selector() {
             </tbody>
          </table>
       </div></body></html>"#;
-    let inlined = inline(&html).unwrap();
+    let inlined = inline(html).unwrap();
     assert_eq!(
         inlined,
         r#"<html><head>
       <title>Test</title>
-      <style>.parent {
-         overflow: hidden;
-         box-shadow: 0 4px 10px 0px rgba(0, 0, 0, 0.1);
-         }
-         .parent > table > tbody > tr > td,
-         .parent > table > tbody > tr > td > div {
-         border-radius: 3px;
-         }
-      </style>
+      
    </head>
    <body>
       <div class="parent" style="overflow: hidden;box-shadow: 0 4px 10px 0px rgba(0, 0, 0, 0.1);">
@@ -347,8 +333,7 @@ fn invalid_rule() {
 #[test]
 fn remove_style_tag() {
     let html = html!("h1 {background-color: blue;}", "<h1>Hello world!</h1>");
-    let inliner = CSSInliner::compact();
-    let result = inliner.inline(&html).unwrap();
+    let result = inline(&html).unwrap();
     assert_eq!(result, "<html><head><title>Test</title></head><body><h1 style=\"background-color: blue;\">Hello world!</h1></body></html>")
 }
 
@@ -377,8 +362,7 @@ a {
 </body>
 </html>
     "#;
-    let inliner = CSSInliner::compact();
-    let result = inliner.inline(html).unwrap();
+    let result = inline(html).unwrap();
     assert_eq!(
         result,
         r#"<html><head>
@@ -450,7 +434,7 @@ fn do_not_process_style_tag() {
     let result = inliner.inline(&html).unwrap();
     assert_eq!(
         result,
-        "<html><head><title>Test</title><style>h1 {background-color: blue;}</style></head><body><h1>Hello world!</h1></body></html>"
+        "<html><head><title>Test</title></head><body><h1>Hello world!</h1></body></html>"
     )
 }
 
@@ -482,7 +466,7 @@ fn extra_css() {
     let result = inliner.inline(&html).unwrap();
     assert_eq!(
         result,
-        "<html><head><title>Test</title><style>h1 {background-color: blue;}</style></head><body><h1 style=\"background-color: green;\">Hello world!</h1></body></html>"
+        "<html><head><title>Test</title></head><body><h1 style=\"background-color: green;\">Hello world!</h1></body></html>"
     )
 }
 
@@ -696,7 +680,7 @@ fn customize_inliner() {
         ..Default::default()
     };
     assert!(!options.load_remote_stylesheets);
-    assert!(!options.remove_style_tags);
+    assert!(options.remove_style_tags);
     assert_eq!(options.base_url, None);
 }
 
@@ -717,5 +701,5 @@ fn inline_to() {
     let html = html!("h1 { color: blue }", r#"<h1>Big Text</h1>"#);
     let mut out = Vec::new();
     css_inline::inline_to(&html, &mut out).unwrap();
-    assert_eq!(String::from_utf8_lossy(&out), "<html><head><title>Test</title><style>h1 { color: blue }</style></head><body><h1 style=\"color: blue ;\">Big Text</h1></body></html>")
+    assert_eq!(String::from_utf8_lossy(&out), "<html><head><title>Test</title></head><body><h1 style=\"color: blue ;\">Big Text</h1></body></html>")
 }