+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Set `remove_style_tags` to `true` by default.`
+- Replaced `--remove-style-tags` with `--keep-style-tags` in CLI.
+
 ## [0.9.0] - 2023-06-10
 
 ### Fixed